refactor: modularize open-sse/services/combo.ts (Issue #3594)
- Splits the 4,600-line combo.ts monolithic file into 12 domain-specific files
- Extracts Auto-Combo, Round-Robin, Quota tracking, Shadow routing, and Context Affinity logic into dedicated modules
- Uses folder-as-module pattern with index.ts facade for zero backward-compatibility breakages
- Auto-prunes and manages all imports for strict typechecking
